# Add native handling of `unspecified()`