--- a/src/cast.c
+++ b/src/cast.c
@@ -271,6 +271,13 @@ static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_ar
   enum vctrs_type x_type = vec_typeof(x);
   enum vctrs_type to_type = vec_typeof(to);
 
+  if (x_type == vctrs_type_scalar) {
+    stop_scalar_type(x, x_arg);
+  }
+  if (to_type == vctrs_type_scalar) {
+    stop_scalar_type(to, to_arg);
+  }
+
   if (x_type == vctrs_type_unspecified) {
     return vec_init(to, vec_size(x));
   }