--- a/src/conditions.c
+++ b/src/conditions.c
@@ -66,6 +66,31 @@ void stop_recycle_incompatible_size(R_len_t x_size, R_len_t size,
   Rf_error("Internal error: `stop_recycle_incompatible_size()` should have jumped earlier");
 }
 
+void stop_incompatible_type(SEXP x,
+                            SEXP y,
+                            struct vctrs_arg* x_arg,
+                            struct vctrs_arg* y_arg) {
+  SEXP syms[5] = {
+    syms_x,
+    syms_y,
+    r_sym("x_arg"),
+    r_sym("y_arg"),
+    NULL
+  };
+  SEXP args[5] = {
+    PROTECT(r_protect(x)),
+    PROTECT(r_protect(y)),
+    PROTECT(vctrs_arg(x_arg)),
+    PROTECT(vctrs_arg(y_arg)),
+    NULL
+  };
+
+  SEXP call = PROTECT(r_call(r_sym("stop_incompatible_type"), syms, args));
+  Rf_eval(call, vctrs_ns_env);
+
+  Rf_error("Internal error: `stop_incompatible_type()` should have jumped earlier");
+}
+
 void stop_corrupt_factor_levels(SEXP x, struct vctrs_arg* arg) {
   SEXP call = PROTECT(Rf_lang3(Rf_install("stop_corrupt_factor_levels"),
                                PROTECT(r_protect(x)),