--- a/src/type-info.c
+++ b/src/type-info.c
@@ -119,6 +119,12 @@ static bool class_is_null(SEXP x) {
 
 // [[ include("vctrs.h") ]]
 enum vctrs_type vec_typeof(SEXP x) {
+  // Check for unspecified objects created with `unspecified()`, which
+  // are objects, meaning they don't go through `vec_base_typeof()`
+  if (vec_is_unspecified(x)) {
+    return vctrs_type_unspecified;
+  }
+
   if (!OBJECT(x) || class_is_null(x)) {
     return vec_base_typeof(x, false);
   }
@@ -154,17 +160,18 @@ void vctrs_stop_unsupported_type(enum vctrs_type type, const char* fn) {
 
 const char* vec_type_as_str(enum vctrs_type type) {
   switch (type) {
-  case vctrs_type_null:      return "null";
-  case vctrs_type_logical:   return "logical";
-  case vctrs_type_integer:   return "integer";
-  case vctrs_type_double:    return "double";
-  case vctrs_type_complex:   return "complex";
-  case vctrs_type_character: return "character";
-  case vctrs_type_raw:       return "raw";
-  case vctrs_type_list:      return "list";
-  case vctrs_type_dataframe: return "dataframe";
-  case vctrs_type_s3:        return "s3";
-  case vctrs_type_scalar:    return "scalar";
+  case vctrs_type_null:         return "null";
+  case vctrs_type_logical:      return "logical";
+  case vctrs_type_integer:      return "integer";
+  case vctrs_type_double:       return "double";
+  case vctrs_type_complex:      return "complex";
+  case vctrs_type_character:    return "character";
+  case vctrs_type_raw:          return "raw";
+  case vctrs_type_list:         return "list";
+  case vctrs_type_dataframe:    return "dataframe";
+  case vctrs_type_s3:           return "s3";
+  case vctrs_type_unspecified:  return "unspecified";
+  case vctrs_type_scalar:       return "scalar";
   }
   never_reached("vec_type_as_str");
 }