--- a/src/type.c
+++ b/src/type.c
@@ -8,23 +8,23 @@ static SEXP fns_vec_type_finalise_dispatch = NULL;
 
 
 static SEXP vec_type_slice(SEXP x, SEXP empty);
-static SEXP lgl_type(SEXP x);
 static SEXP s3_type(SEXP x);
 
 // [[ include("vctrs.h"); register() ]]
 SEXP vec_type(SEXP x) {
   switch (vec_typeof(x)) {
-  case vctrs_type_scalar:    return x;
-  case vctrs_type_null:      return R_NilValue;
-  case vctrs_type_logical:   return lgl_type(x);
-  case vctrs_type_integer:   return vec_type_slice(x, vctrs_shared_empty_int);
-  case vctrs_type_double:    return vec_type_slice(x, vctrs_shared_empty_dbl);
-  case vctrs_type_complex:   return vec_type_slice(x, vctrs_shared_empty_cpl);
-  case vctrs_type_character: return vec_type_slice(x, vctrs_shared_empty_chr);
-  case vctrs_type_raw:       return vec_type_slice(x, vctrs_shared_empty_raw);
-  case vctrs_type_list:      return vec_type_slice(x, vctrs_shared_empty_list);
-  case vctrs_type_dataframe: return df_map(x, &vec_type);
-  case vctrs_type_s3:        return s3_type(x);
+  case vctrs_type_scalar:      return x;
+  case vctrs_type_null:        return R_NilValue;
+  case vctrs_type_logical:     return vec_type_slice(x, vctrs_shared_empty_lgl);
+  case vctrs_type_integer:     return vec_type_slice(x, vctrs_shared_empty_int);
+  case vctrs_type_double:      return vec_type_slice(x, vctrs_shared_empty_dbl);
+  case vctrs_type_complex:     return vec_type_slice(x, vctrs_shared_empty_cpl);
+  case vctrs_type_character:   return vec_type_slice(x, vctrs_shared_empty_chr);
+  case vctrs_type_raw:         return vec_type_slice(x, vctrs_shared_empty_raw);
+  case vctrs_type_list:        return vec_type_slice(x, vctrs_shared_empty_list);
+  case vctrs_type_dataframe:   return df_map(x, &vec_type);
+  case vctrs_type_s3:          return s3_type(x);
+  case vctrs_type_unspecified: return vctrs_shared_empty_uns;
   }
   never_reached("vec_type_impl");
 }
@@ -35,13 +35,6 @@ static SEXP vec_type_slice(SEXP x, SEXP empty) {
   } else {
     // Slicing preserves attributes
     return vec_slice(x, R_NilValue);
-  }
-}
-static SEXP lgl_type(SEXP x) {
-  if (vec_is_unspecified(x)) {
-    return vctrs_shared_empty_uns;
-  } else {
-    return vec_type_slice(x, vctrs_shared_empty_lgl);
   }
 }
 static SEXP s3_type(SEXP x) {