… src/typeof2-s3.c
… src/typeof2.c
… src/vctrs.h
… tests/testthat/test-cast.R
… tests/testthat/test-compare.R
… tests/testthat/test-equal.R
… tests/testthat/test-size.R
… tests/testthat/test-slice-assign.R
… tests/testthat/test-slice.R
… tests/testthat/test-type-bare.R
… tests/testthat/test-type-unspecified.R
… tests/testthat/test-type2.R