This PR adds native handling of `unspecified()`, which is one of the things slowing down https://github.com/tidyverse/dplyr/issues/4561.

This PR is the result of much conversation with @lionel- about what unspecified should be doing, especially in relation to `vec_ptype2()` and `vec_ptype_common()`.

This PR adds 5 major things:

1. `vec_typeof()` can now return `vctrs_type_unspecified`. This is helpful for controlling ptype2 and cast methods. `vec_base_typeof()` will never return this, meaning that `vec_proxy_typeof()` will continue to return `vctrs_type_logical` for unspecified vectors. This is nice because this is what `vec_compare()` and `vec_equal()` end up using, so no additional special handling is required for this new internal type.

2. The cast dispatch works by catching if `x` is unspecified, and having special behavior if so. Most of the time it just calls `vec_init(to, vec_size(x))`. If `y` happens to be s3, then a dispatch routine specific to `x` being unspecified and `y` being s3 is run to allow us to handle any "known" s3 classes more efficiently.

3. The ptype2 dispatch works similarly to the cast dispatch, except it checks if either `x` or `y` is unspecified. There is special handling for `vec_ptype2(NA, list())` here, which _currently_ should return an error, although @lionel- and I are unsure of this.

4. `vec_typeof2()` and `vec_typeof2_s3()` know about `vctrs_type_unspecified`. Even though the enum values aren't really used in the ptype2 and cast dispatch, I think it makes sense to flesh these out.

5. After talking with @lionel-, I've changed `vec_ptype2(NA, NA)` so that it returns `unspecified()`. `vec_ptype_common()` will finalise this, so it will still return `logical()`. This fixes iterative `vec_ptype2()` calls like `vec_ptype2(vec_ptype2(NA, NA), "")`, which previously errored.

Below are some benchmarks. Improvements are pretty good, but I'd like to see more improvements in the `vec_ptype2(<unspecified>, <bare-factor>)` cases. This will end up calling `vec_type()`, which for S3 objects calls `vec_slice(<bare-factor>, NULL)`. An improved factor slicing method might help here, but I'm not sure yet. We could also special case our "known" classes in the `s3_type()` routine that `vec_type()` calls, avoiding the slice entirely.

``` r
library(vctrs)

na <- NA
uns <- unspecified()

x <- 1
y <- factor("x")
```

```r
# before
bench::mark(vec_ptype2(uns, uns))
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, uns)   12.4µs     16µs    52382.    15.1KB     15.7

# after
bench::mark(vec_ptype2(uns, uns))
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, uns)    945ns   1.23µs   564630.        0B        0
```

```r
# before
bench::mark(vec_ptype2(na, y))
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(na, y)   12.7µs   16.3µs    60953.    26.7KB     24.4

# after
bench::mark(vec_ptype2(na, y))
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(na, y)   7.79µs   10.3µs    91157.    10.7KB     18.2
```

```r
# before
bench::mark(vec_ptype2(uns, x))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, x)   3.78µs   5.07µs   194677.        0B     19.5

# after
bench::mark(vec_ptype2(uns, x))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, x)    758ns   1.01µs   866591.        0B        0
```

```r
# before
bench::mark(vec_ptype2(uns, y))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, y)   10.7µs   12.7µs    73955.        0B     22.2

# after
bench::mark(vec_ptype2(uns, y))
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(uns, y)   7.37µs   8.44µs   113308.        0B     34.0
```

```r
# before
bench::mark(vec_cast(na, y))
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(NA, y)   16.7µs   19.7µs    47827.    33.4KB     23.9

# after
bench::mark(vec_cast(na, y))
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(na, y)   5.89µs   6.88µs   133945.        0B     26.8
```

```r
# before
bench::mark(vec_cast(uns, x))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(uns, x)   13.5µs   15.6µs    61904.    7.01KB     24.8

# after
bench::mark(vec_cast(uns, x))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(uns, x)   1.19µs   1.66µs   590555.        0B        0
```

```r
# before
bench::mark(vec_cast(uns, y))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(uns, y)   18.9µs   21.6µs    44425.        0B     17.8

# after
bench::mark(vec_cast(uns, y))
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(uns, y)   6.08µs   7.18µs   132693.        0B     26.5
```